Keep cover color cache guards testable
